마이페이지 즐겨찾기 기능구현 오류 수정

+++ s20230404/src/main/java/com/travelAlone/s20230404/dao/km/MypageDaoImpl.java
@@ -134,7 +134,7 @@ public class MypageDaoImpl implements MypageDao{
 
     @Override
     public int kmMypageFavoritesCountRes(Long id) {
-        return session.selectOne("kmMypageFavoritesCountRes",id);
+        return session.selectOne("kmMypageFavoritesCountRes",(long)id);
     }
 
     @Override
